+++ lib/src/network/winr_api.dart
@@ -284,7 +284,7 @@ class WINRRequestDefaults {
   WINRRequestDefaults._();
 
   static String platformOS = 'iOS';
-  static String sdkVersion = '2.9.4';
+  static String sdkVersion = '2.9.5';
 }
 
 /// Response from claiming daily entries (mirrors iOS